requirements.txt
twitchio==2.1.0
PyYAML==6.0